add labels & move function
